chore(toolpack): move swc position
