+++ packages/toolpack/src/webpack/loaders/shuvi-swc-loader/__tests__/core.test.ts
