+++ packages/toolpack/src/webpack/loaders/shuvi-swc-loader/__tests__/plugins/dynamic-plugin.test.ts
