+++ packages/toolpack/src/webpack/loaders/shuvi-swc-loader/__tests__/plugins/jsx-pragma.test.ts
