+++ packages/toolpack/src/webpack/loaders/shuvi-swc-loader/__tests__/plugins/react-remove-properties.test.ts
