+++ packages/toolpack/src/webpack/loaders/shuvi-swc-loader/__tests__/plugins/remove-console.test.ts
