+++ packages/toolpack/src/webpack/loaders/shuvi-swc-loader/load-sources.js
@@ -4,14 +4,13 @@ const path = require('path');
 const { platformArchTriples } = require('@napi-rs/triples');
 const ArchName = arch();
 const PlatformName = platform();
-
 let bindings;
 let loadError;
 
 const triples = platformArchTriples[PlatformName][ArchName];
 for (const triple of triples) {
   try {
-    const swcSource = path.join(__dirname, '../../swc-source');
+    const swcSource = path.join(__dirname, '../../../../swc-source');
     const localFilePath = path.join(
       swcSource,
       `shuvi-swc.${triple.platformArchABI}.node`
